# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kindue household bill coordination API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /households:
    get:
      operationId: listHouseholds
      tags: [households]
      summary: List households the current user belongs to
      responses:
        "200":
          description: List of households
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Household"
    post:
      operationId: createHousehold
      tags: [households]
      summary: Create a new household
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHouseholdBody"
      responses:
        "201":
          description: Created household
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
  /households/{householdId}:
    get:
      operationId: getHousehold
      tags: [households]
      summary: Get a household
      parameters:
        - $ref: "#/components/parameters/householdId"
      responses:
        "200":
          description: Household
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateHousehold
      tags: [households]
      summary: Update household details
      parameters:
        - $ref: "#/components/parameters/householdId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHouseholdBody"
      responses:
        "200":
          description: Updated household
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
    delete:
      operationId: deleteHousehold
      tags: [households]
      summary: Delete a household (PU only)
      parameters:
        - $ref: "#/components/parameters/householdId"
      responses:
        "204":
          description: Deleted
  /households/{householdId}/dashboard:
    get:
      operationId: getHouseholdDashboard
      tags: [households]
      summary: Get dashboard summary for a household
      parameters:
        - $ref: "#/components/parameters/householdId"
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  # ── MEMBERS ──────────────────────────────────────────────────────────────────
  /households/{householdId}/members:
    get:
      operationId: listHouseholdMembers
      tags: [members]
      summary: List members of a household
      parameters:
        - $ref: "#/components/parameters/householdId"
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HouseholdMember"
    post:
      operationId: inviteHouseholdMember
      tags: [members]
      summary: Invite a member to the household (PU/Trustee only)
      parameters:
        - $ref: "#/components/parameters/householdId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberBody"
      responses:
        "201":
          description: Invited member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HouseholdMember"
  /households/{householdId}/members/{memberId}:
    patch:
      operationId: updateHouseholdMember
      tags: [members]
      summary: Update a member's role (Trustee only, requires re-auth token)
      parameters:
        - $ref: "#/components/parameters/householdId"
        - $ref: "#/components/parameters/memberId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberBody"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HouseholdMember"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: removeHouseholdMember
      tags: [members]
      summary: Remove a member from the household (PU/Trustee only)
      parameters:
        - $ref: "#/components/parameters/householdId"
        - $ref: "#/components/parameters/memberId"
      responses:
        "204":
          description: Removed

  # ── BILLS ────────────────────────────────────────────────────────────────────
...
  /households/{householdId}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for a household
      parameters:
        - $ref: "#/components/parameters/householdId"
        - name: billId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document (all roles can upload)
      parameters:
        - $ref: "#/components/parameters/householdId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDocumentBody"
      responses:
        "201":
          description: Uploaded document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /households/{householdId}/documents/{documentId}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document metadata
      parameters:
        - $ref: "#/components/parameters/householdId"
        - $ref: "#/components/parameters/documentId"
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document (PU/Trustee only)
      parameters:
        - $ref: "#/components/parameters/householdId"
        - $ref: "#/components/parameters/documentId"
      responses:
        "204":
          description: Deleted

  # ── RECEIPTS ─────────────────────────────────────────────────────────────────
  /households/{householdId}/bills/{billId}/receipts:
    get:
      operationId: listReceipts
      tags: [documents]
      summary: List receipts for a bill
      parameters:
        - $ref: "#/components/parameters/householdId"
        - $ref: "#/components/parameters/billId"
      responses:
        "200":
          description: Receipts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadReceipt
      tags: [documents]
      summary: Upload a payment receipt for a bill (required for CG/Other to mark paid)
      parameters:
        - $ref: "#/components/parameters/householdId"
        - $ref: "#/components/parameters/billId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDocumentBody"
      responses:
        "201":
          description: Uploaded receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  # ── TRIAGE ───────────────────────────────────────────────────────────────────
...
components:
  parameters:
    householdId:
      name: householdId
      in: path
      required: true
      schema:
        type: integer
    memberId:
      name: memberId
      in: path
      required: true
      schema:
        type: integer
    billId:
      name: billId
      in: path
      required: true
      schema:
        type: integer
    documentId:
      name: documentId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── USER ────────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - email
        - createdAt
...
    Household:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        primaryUserId:
          type: integer
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - primaryUserId
        - memberCount
        - createdAt
        - updatedAt
    CreateHouseholdBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
      required:
        - name
    UpdateHouseholdBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
    DashboardSummary:
      type: object
      properties:
        overdueCount:
          type: integer
        dueSoonCount:
          type: integer
          description: Bills due within 7 days
        pendingApprovalCount:
          type: integer
        totalMonthlyEstimate:
          type: number
        hasLowBalanceRisk:
          type: boolean
          description: True if any linked account balance is below safe threshold (PU/T only)
        topTriageItems:
          type: array
          items:
            $ref: "#/components/schemas/TriageItem"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEntry"
      required:
        - overdueCount
        - dueSoonCount
        - pendingApprovalCount
        - totalMonthlyEstimate
        - hasLowBalanceRisk
        - topTriageItems
        - recentActivity

    # ── MEMBERS ──────────────────────────────────────────────────────────────────
    HouseholdMember:
      type: object
      properties:
        id:
          type: integer
        householdId:
          type: integer
        userId:
          type: integer
        role:
          type: string
          enum: [primary_user, trustee, caregiver, other]
        inviteEmail:
          type: ["string", "null"]
        inviteStatus:
          type: string
          enum: [pending, accepted, declined]
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - householdId
        - userId
        - role
        - inviteStatus
        - createdAt
        - updatedAt
    InviteMemberBody:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
          enum: [trustee, caregiver, other]
      required:
        - email
        - role
    UpdateMemberBody:
      type: object
      properties:
        role:
          type: string
          enum: [trustee, caregiver, other]
        sensitiveToken:
          type: string
          description: Re-auth token required for role changes
      required:
        - role
        - sensitiveToken

    # ── BILLS ────────────────────────────────────────────────────────────────────
    Bill:
      type: object
      properties:
        id:
          type: integer
        householdId:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [utilities, telecom, insurance, mortgage_rent, auto_loan, credit_card, tolls, subscription, other]
        amount:
          type: ["number", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [due, overdue, paid, pending_approval, rejected]
        isRecurring:
          type: boolean
        recurrenceInterval:
          type: ["string", "null"]
          enum: [weekly, monthly, quarterly, annual, "null"]
        notes:
          type: ["string", "null"]
        billerName:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        requiresReceipt:
          type: boolean
          description: True if the next mark-paid action requires a receipt (for CG/Other)
        receiptUploaded:
          type: boolean
        riskFlags:
          type: array
          items:
            $ref: "#/components/schemas/RiskFlag"
        createdByRole:
          type: string
          enum: [primary_user, trustee, caregiver, other]
        assignedToMemberId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - householdId
        - name
        - category
        - status
        - isRecurring
        - requiresReceipt
        - receiptUploaded
        - riskFlags
        - createdByRole
        - createdAt
        - updatedAt
...
    RiskFlag:
      type: object
      properties:
        type:
          type: string
          enum: [overdue, shutoff_risk, late_fee_risk, low_balance_risk, upcoming_due]
        severity:
          type: string
          enum: [low, medium, high, critical]
        message:
          type: string
        daysUntilDue:
          type: ["integer", "null"]
      required:
        - type
        - severity
        - message

    # ── TRIAGE ───────────────────────────────────────────────────────────────────
    TriageItem:
      type: object
      properties:
        bill:
          $ref: "#/components/schemas/Bill"
        priorityScore:
          type: number
          description: Computed priority (higher = handle first)
        priorityReason:
          type: string
          description: Human-readable explanation of the ranking
        rank:
          type: integer
      required:
        - bill
        - priorityScore
        - priorityReason
        - rank
...
    Document:
      type: object
      properties:
        id:
          type: integer
        householdId:
          type: integer
        billId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [statement, receipt, other]
        fileName:
          type: string
        fileSize:
          type: integer
        mimeType:
          type: string
        storageKey:
          type: string
        downloadUrl:
          type: string
        uploadedByUserId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - householdId
        - type
        - fileName
        - fileSize
        - mimeType
        - storageKey
        - downloadUrl
        - uploadedByUserId
        - createdAt
    UploadDocumentBody:
      type: object
      properties:
        fileName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        storageKey:
          type: string
          description: Object storage key from direct upload
        billId:
          type: integer
        type:
          type: string
          enum: [statement, receipt, other]
      required:
        - fileName
        - mimeType
        - fileSize
        - storageKey
        - type

    # ── GMAIL ─────────────────────────────────────────────────────────────────────
...
    AuditLogEntry:
      type: object
      properties:
        id:
          type: integer
        householdId:
          type: integer
        actorUserId:
          type: integer
        actorName:
          type: ["string", "null"]
        action:
          type: string
          description: e.g. bill.approved, member.role_changed, bill.paid
        entityType:
          type: string
          description: e.g. bill, member, document
        entityId:
          type: ["integer", "null"]
        details:
          type: ["string", "null"]
          description: Human-readable summary of the change
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - householdId
        - actorUserId
        - action
        - entityType
        - createdAt

    # ── NOTIFICATIONS ─────────────────────────────────────────────────────────────
...